refactor(llm): split admin commands into commands/admin.tsx subplugin
Move llm.providers / llm.models / llm.reset / llm.stop / llm.catalog /
llm.memory into a dedicated BasePlugin subclass loaded via ctx.plugin().
Subplugin lifecycle is owned by koishi: when PluginLLM unloads, koishi
disposes the subplugin too and unregisters its commands automatically.

The subplugin is stateless — it reads everything from ctx.llm (parent
plugin instance), so all the existing service refs (memory, memoryFork,
activeChats, catalog, providers, config) flow through naturally.

resolveMemoryKey is bumped from private to public so subplugins can
share that platform/userId derivation.

Via Claude Code
